Revision from review comments
